src/ons_api/requirements.txt
fastapi
uvicorn
pandas
pyarrow
google-cloud-storage
requests